refactored File Structure and naming conventions
