--- src/components/navbar/Navbar.tsx
@@ -1,36 +1,0 @@
-import React, { SetStateAction } from "react";
-import { Link } from "react-router-dom";
-import { navLinks } from "./NavbarLinks";
-import MenuIcon from "@mui/icons-material/Menu";
-
-type NavbarProps = {
-  setIsOpen: React.Dispatch<SetStateAction<boolean>>;
-};
-
-const Navbar = ({ setIsOpen }: NavbarProps) => {
-  return (
-    <div className="h-[80px] border-b flex text-xl shadow-md">
-      <div className=" flex w-full justify-between self-center">
-        <div onClick={() => setIsOpen((prev: boolean) => !prev)} className="ml-5">
-          <span>
-            <MenuIcon />
-          </span>
-        </div>
-        <div className="flex justify-evenly w-full self-center">
-          <Link to={"/dashboard"} className="sm:hidden xl:flex">
-            logo
-          </Link>
-          {navLinks.map((el) => (
-            <div key={el.path} className="sm:flex hidden">
-              <div className="flex self-center">
-                <Link to={el.path}>{el.element}</Link>
-              </div>
-            </div>
-          ))}
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Navbar;
