--- src/components/navbar/NavbarLinks.ts
@@ -1,17 +1,0 @@
-  export const navLinks = [
-        {
-            path: "/admin", element: "Panel admina",
-        },
-        {
-            path: "/notifications", element: "Powiadomienia",
-        },
-        {
-            path: "/menu", element: "Menu",
-        },
-        {
-            path: "/raports", element: "Raporty",
-        },
-        {
-            path: "/settings", element: "Ustawienia",
-        }
-    ];
