--- src/components/sidebar/content/SidebarLinks.tsx
@@ -1,16 +1,0 @@
-const SidebarLinks = [
-  {
-    main: "Magazyn",
-    link: "",
-  },
-  {
-    main: "Magazyn",
-    link: "",
-  },
-  {
-    main: "Magazyn",
-    link: "",
-  },
-];
-
-export default SidebarLinks;
