--- src/components/userSettingsComponent/UserPersonalSettingsOption.tsx
@@ -1,41 +1,0 @@
-import { SetStateAction } from "react";
-import { handleChangeInput } from "../../formhandleChangeLogic/HandlingChangeInput";
-import { UserPersonalData } from "./UserSettings";
-import { UserSettingsContent } from "./content/UserSettingsContent";
-import React from "react";
-
-type UserPersonalSettingsOptionProps = {
-  userSettingsData: UserPersonalData;
-  setUserSettingsData: React.Dispatch<SetStateAction<UserPersonalData>>;
-  userSettings: UserSettingsContent[];
-};
-
-export const UserPersonalSettingsOption = ({
-  userSettingsData,
-  setUserSettingsData,
-  userSettings,
-}: UserPersonalSettingsOptionProps) => {
-  return (
-    <div className="mt-8 p-6 bg-white rounded-md shadow-md">
-      <h2 className="text-2xl font-semibold mb-4">Ustawienia Użytkownika</h2>
-      <div className="flex flex-col">
-        {userSettings.map((settings) => (
-          <div key={settings.key} className="mb-4">
-            <label className="block text-sm font-medium text-gray-600">{settings.label}</label>
-            <div className="flex">
-              <input
-                type="text"
-                className="mt-1 p-2 w-full border rounded-md"
-                value={userSettingsData[settings.key as keyof UserPersonalData]}
-                onChange={(e) => handleChangeInput(setUserSettingsData, e, settings)}
-              />
-            </div>
-          </div>
-        ))}
-      </div>
-      <button className="flex w-full justify-center bg-blue-500 text-white px-4 py-2 rounded-md hover:bg-blue-600">
-        Zapisz Ustawienia
-      </button>
-    </div>
-  );
-};
