+++ src/data/adminManagmentOptionsData/AminManagmentOptionsPickerData.ts
@@ -1,0 +1,16 @@
+import { type TAdminManagmentOptionsData } from "../dataTypes/data-types-d";
+
+export const adminManagmentOptionsPickerData: TAdminManagmentOptionsData[] = [
+    {
+        text: "Zarządzanie Użytkownikami",
+        key: "employees",
+    },
+    {
+        text: "Zarządzanie Firmą",
+        key: "company",
+    },
+    {
+        text: "Zarządzanie Dochodami",
+        key: "market",
+    },
+];
