+++ src/data/adminManagmentTableData/AdminManagmentUserData.ts
@@ -1,0 +1,70 @@
+export const userTableHeaders = [
+    { field: "id", headerName: "ID", width: 90 },
+    {
+        field: "firstName",
+        headerName: "First name",
+        width: 150,
+        editable: true,
+    },
+    {
+        field: "lastName",
+        headerName: "Last name",
+        width: 150,
+        editable: true,
+    },
+    {
+        field: "age",
+        headerName: "Age",
+        type: "number",
+        width: 110,
+        editable: true,
+    },
+    {
+        field: "fullName",
+        headerName: "Full name",
+        description: "This column has a value getter and is not sortable.",
+        sortable: false,
+        width: 160,
+        valueGetter: (params: any) => `${params.row.firstName || ""} ${params.row.lastName || ""}`, //TODO change type of params
+    },
+    {
+        field: "phoneNumber",
+        headerName: "Phone number",
+        type: "number",
+        sortable: true,
+        width: 150,
+    },
+    {
+        field: "city",
+        headerName: "City",
+        type: "text",
+        sortable: true,
+        width: 150,
+    },
+    {
+        field: "state",
+        headerName: "State",
+        type: "text",
+        sortable: true,
+        width: 150,
+    },
+    {
+        field: "country",
+        headerName: "Country",
+        type: "text",
+        sortable: true,
+        width: 150,
+    },
+];
+
+export const userData = [
+    { id: 1, lastName: "Snow", firstName: "Jon", age: 35 },
+    { id: 2, lastName: "Lannister", firstName: "Cersei", age: 42 },
+    { id: 3, lastName: "Lannister", firstName: "Jaime", age: 45 },
+    { id: 4, lastName: "Stark", firstName: "Arya", age: 16 },
+    { id: 5, lastName: "Targaryen", firstName: "Daenerys", age: null },
+    { id: 6, lastName: "Melisandre", firstName: null, age: 150 },
+    { id: 7, lastName: "Clifford", firstName: "Ferrara", age: 44 },
+    { id: 8, lastName: "Frances", firstName: "Rossini", age: 36 },
+    { id: 9, lastName: "Roxie", firstName: "Harvey", age: 65 },
+];
