+++ src/data/dataTypes/data-types-d.ts
@@ -1,0 +1,16 @@
+import { SetStateAction } from "react";
+
+// 
+export type TAdminManagmentOptionsData = {
+  text: string;
+  key: string;
+};
+
+export type TAdminSelectedOptionProps = {
+  selectedOptionByAdmin: string;
+  setSelectedOptionByAdmin: React.Dispatch<SetStateAction<string>>;
+};
+
+export type Options = {
+  option: string;
+};
