+++ src/data/navbarStaticData/NavbarNavigationLinks.ts
@@ -1,0 +1,17 @@
+export const navLinks = [
+    {
+        path: "/admin", element: "Panel admina",
+    },
+    {
+        path: "/notifications", element: "Powiadomienia",
+    },
+    {
+        path: "/menu", element: "Menu",
+    },
+    {
+        path: "/raports", element: "Raporty",
+    },
+    {
+        path: "/settings", element: "Ustawienia",
+    }
+];
