+++ src/data/sidebarStaticData/SidebarNavigationLinks.ts
@@ -1,0 +1,14 @@
+export const SidebarLinks = [
+    {
+        path: "",
+        element: "EXAMPLE",
+    },
+    {
+        path: "",
+        element: "EXAMPLE",
+    },
+    {
+        path: "",
+        element: "EXAMPLE",
+    },
+];
