+++ src/pages/protectedRoutes/UserSettingsPage.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import UserSettings from "../../features/userSettings/UserSettings";
+
+const UserSettingsPage = () => {
+    return (
+        <div className='w-full'>
+            <UserSettings/>
+        </div>
+    );
+};
+
+export default UserSettingsPage;
