+++ src/pages/publicRoutes/RegisterPage.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import Register from "../../features/authentication/Register";
+
+const RegisterPage = () => {
+  return (
+    <div className="h-screen flex justify-center">
+      <form className="flex flex-col self-center">
+        <Register />
+      </form>
+    </div>
+  );
+};
+
+export default RegisterPage;
