--- src/pages/routes/AdminPanelPage.tsx
@@ -1,12 +1,0 @@
-import React from "react";
-import AdminPanel from "../../components/adminPanel/AdminPanel";
-
-const AdminPanelPage = () => {
-  return (
-    <div>
-      <AdminPanel />
-    </div>
-  );
-};
-
-export default AdminPanelPage;
