--- src/pages/routes/UserSettingsPage.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import UserSettings from "../../components/userSettingsComponent/UserSettings";
-
-const UserSettingsPage = () => {
-    return (
-        <div className='w-full'>
-            <UserSettings/>
-        </div>
-    );
-};
-
-export default UserSettingsPage;
