+++ src/utils/RegisterAuthentication.ts
@@ -1,0 +1,13 @@
+import axios from "axios";
+import { TRegisterCredentials } from "../data/dataTypes/user-creds-types-d";
+
+const HandleRegisterCompany = async (formData: TRegisterCredentials) => {
+  const ApiUrl = process.env.REACT_APP_URL + "/Auth/Register";
+  try {
+    await axios.post(ApiUrl, formData);
+  } catch (error) {
+    console.error(error)
+  }
+};
+
+export default HandleRegisterCompany
